Remove external tools deps and make safer
Running system()-calls to 'sed' is not portable. Some systems have
BSD-sed, and others GNU-sed. These do not accept similar options.
Replaced the system calls with ruby-equivalentish (not stream-editing,
but good enough for this purpose since we control the input files sizes.

Also added a safeguard on the find(). Now we make sure that the
find()-command is run in the expected directory with __FILE__ and not
working-directory, which can be anything.